transfer MemoryAllocator to LineReader; increase direct memory size for reduce memory copying in compact method
